add context info to logs using a callback
This change introduces a new capability to add context information into
every log message using a block

This is inteneded to be used so that we can inject trace and span
context information from an o11y library into our log messages so that
we can correlate between traces and logs.

Currently, when adding this information into logs we are having to wrap
the Twiglet::Logger in a delegator or subclass which reimplements some
of its interface and internals. Each team has done this in a different
way, which will lead to a long term maintenance problem so providing an API
for this will hopefully reduce our maintenance costs.

Note: `.with` doesn't provide this behaviour as we need to change the context
information on every log event. It also allocates a new logger, which is
unnecessary.