preserve context when adding a new context provider
This is common need when creating a domain specific logger: we may want to provide domain specific ids (like user/session ids) inline with some generic ones (like Otel ids). Currently, we have to create a single lambda with all of the behaviour in it, which leads to lots of duplicated across apps.